refactor: split db auth.rs into focused modules (#202)
Mechanically split the 1,431-line db/src/auth.rs into a thin module root
(error/domain types + cross-cutting row_to_user/now_unix helpers + wide
re-export surface + test_support) plus seven submodules under db/src/auth/:
password, token, users, login, session, device, session_key. Tests distributed
next to their target functions. The full external API (omnibus_db::auth::*,
consumed across server/ and frontend/) is preserved via re-exports. No behavior
change.
